requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
astunparse==1.6.3
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.1.8
deepface==0.0.95
exceptiongroup==1.3.0
fastapi==0.118.2
filelock==3.19.1
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.75.1
gTTS==2.5.4
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.10.0
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
mtcnn==0.1.1
namex==0.1.0
numpy==2.0.2
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.32.1
psycopg2-binary==2.9.9
pydantic==2.12.0
pydantic_core==2.41.1
pygame==2.6.1
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.5
retina-face==0.0.17
rich==14.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tf_keras==2.20.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
Werkzeug==3.1.3
wrapt==1.17.3
zipp==3.23.0